% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.area.collists.R
\name{make.area.collLists}
\alias{make.area.collLists}
\title{Make collection lists for each area}
\usage{
make.area.collLists(
  x,
  fileHeader = "",
  outdir = "tablesByArea.snapshot.\%Y-\%m-\%d_\%H.\%M",
  verbose = TRUE,
  area.list = NA,
  do.all = FALSE,
  ...
)
}
\description{
Make collection lists for each area
}
