% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.mor.R
\name{read.mor}
\alias{read.mor}
\title{Read in collections lists data}
\usage{
read.mor(directory = getwd())
}
\arguments{
\item{directory}{Working directory where files are located}
}
\description{
Reads in the tables for making collections lists, packages them up nice.
Searches for tables with these strings in their names... FINISH DOCS.
}
