% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.collList.R
\name{make.collList}
\alias{make.collList}
\title{Make collections lists}
\usage{
make.collList(
  x,
  excludeYes = TRUE,
  excludeAreas = x$bg.table$bglocation[!x$bg.table$INCLUDE],
  makeSnapshot = TRUE,
  sortTables = TRUE,
  trimL = 40,
  sciQuery = NA,
  famQuery = NA,
  orderQuery = NA,
  queryIgnoreCase = FALSE,
  phenoAdd = TRUE,
  useCollSheets = FALSE,
  makeAccReport = TRUE,
  outdir = format(Sys.time(), "table.snapshot.\%Y-\%m-\%d_\%H.\%M"),
  reportOnlyAccessionsWeCollect = TRUE,
  cleanFamCol = TRUE,
  frFams = c("Cupressaceae", "Pinaceae", "Cephalotaxaceae", "Taxaceae"),
  ...
)
}
\description{
Make collections lists
}
