% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.collList.R
\name{make.collList}
\alias{make.collList}
\title{Make collections lists}
\usage{
make.collList(
  x,
  excludeYes = TRUE,
  excludeAreas = x$bg.table$bglocation[!x$bg.table$INCLUDE],
  makeSnapshot = TRUE,
  sortTables = TRUE,
  trimL = 40,
  sciQuery = NA,
  famQuery = NA,
  orderQuery = NA,
  queryIgnoreCase = FALSE,
  phenoAdd = TRUE,
  useCollSheets = FALSE,
  makeAccReport = TRUE,
  outdir = format(Sys.time(), "table.snapshot.\%Y-\%m-\%d_\%H.\%M"),
  reportOnlyAccessionsWeCollect = TRUE,
  cleanFamCol = TRUE,
  frFams = c("Cupressaceae", "Pinaceae", "Cephalotaxaceae", "Taxaceae"),
  ...
)
}
\description{
This is the master function. Run this on an object from \code{read.mor}
}
\examples{
 \dontRun{
   workingDir <- 'C:/Users/ahipp/Dropbox/Herbarium/BRAHMS/COLL.LISTS.2017/data/AAA.ACTIVE.SHEETS'
   dat <- read.mor(workingDir)
   out <- make.collList(dat, sciQuery = 'Quercus|Acer|Tilia')
 }
}
